% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_frame.R
\name{embedFrame}
\alias{embedFrame}
\title{Embed HTML frames into Jupyter Notebooks}
\usage{
embedFrame(h, height = NULL, width = NULL, scrollable = TRUE)
}
\arguments{
\item{h}{An \code{htmlwidget} object.}

\item{height}{Height of iframe element in pixels (px).}

\item{width}{Width of iframe element in pixels (px).}

\item{scrollable}{Provide a scrollbar if overfill is detected? Defaults to \code{TRUE}.}
}
\description{
This function allows users to override the default height, width, and
scrolling options of \href{https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe}{Inline Frame (iframe)}
elements. This can be used for any general \code{htmlwidgets} objects.
}
